Replace the VAES optimized AES-GCM implementation
The existing VAES-optimized AES-GCM implementation has a binary size of
about 700 KB which makes it not really suitable for production use due
to icache thrashing, and also a source length of almost 5000 lines which
makes it difficult to review.

Replace it with a new implementation that I wrote from scratch.  The
implementation has been incorporated into Linux kernel v6.11; this is
the port of that code to OpenSSL.

This reduces the binary size by about 700 KB and the source code length
by about 3400 lines.  Performance in microbenchmarks is generally about
the same; the real-world performance should be much better due to
improved icache usage.

The new assembly file also supports both 256-bit and 512-bit vectors, so
unlike the old code it is ready for future CPUs that will support
AVX10/256 but not AVX10/512.  However, for now I've only wired up the
512-bit implementation.