Replace the VAES optimized AES-GCM implementation
The existing VAES-optimized AES-GCM implementation has a binary size of
about 700 KB, which makes it not really suitable for production use due
to icache thrashing.  It is also almost 5000 lines of assembly, which
makes it difficult to review.

Replace it with a new implementation that I wrote from scratch.  A
variant of this implementation has been incorporated into Linux kernel
v6.11; this is a port of that code to OpenSSL.  It is modified slightly
from the similar port I did to BoringSSL
(https://boringssl-review.googlesource.com/c/boringssl/+/70187), as
upstream OpenSSL wires up AES-GCM implementations differently.

This reduces the binary size by about 700 KB (over 12% the size of the
entire libcrypto.so) and the source code length by about 3500 lines.

Performance in microbenchmarks is generally comparable.  Meanwhile, the
reduced icache usage should result in more substantial improvements in
practice that are not accounted for in microbenchmarks.

The new assembly file also supports both 256-bit and 512-bit vectors, so
unlike the old code it is ready for any future CPUs that will support
AVX10/256 without AVX10/512.  However, for now I've only wired up the
512-bit implementation for parity with the existing code.